🚀 chore(main.yml): add release workflow
This commit adds a new workflow called "release" to the main.yml file in the .github/workflows directory. The release workflow is triggered after the "test" workflow is completed. It runs on the ubuntu-22.04 runner.

The release workflow includes a step to bump the version and push a tag using the mathieudutour/github-tag-action@v6.0 action. The action requires the `RELEASES_ACTION_GITHUB_TOKEN` secret to be set in the repository's secrets.

This change is made to automate the process of creating releases and tagging versions in the repository.